+++ ts/test/test-utils/index.ts
@@ -1,0 +1,5 @@
+import * as TestUtils from './testUtils';
+import * as Stubs from './stubs';
+export * from './timeout';
+
+export { TestUtils, Stubs };
