--- ts/test/utils/index.ts
@@ -1,3 +1,0 @@
-import * as TestUtils from './testUtils';
-
-export { TestUtils };
